requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
docopt==0.6.2
humanize==0.5.1
idna==2.8
imageio==2.5.0
imutils==0.5.2
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.16.4
opencv-python==4.1.0.25
PCV==1.0
Pillow==6.0.0
pvc==0.3.0
pycollada==0.6
pygame==1.9.6
PyGLM==1.0.0
PyOpenGL==3.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
pythondialog==3.4.0
pyvmomi==6.7.1.2018.12
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
tabulate==0.8.3
tqdm==4.32.2
urllib3==1.25.3
vconnector==0.5.3
